backend/requirements.txt
Flask==2.3.3
flask-socketio==5.3.6
eventlet==0.35.1
flask-cors==4.0.0
python-dotenv==1.0.1